Make use of the alt_id information
Alternative id are now added to the hash as if they were real id, but they point to the normal GO term (we do not create a new one).

So you can call the GO:0006350 but you will retrieve the information of the GO:0006351 which is the normal GO id.